I have decided to scrap and build this project.
The past experiences of the development have pointed out that it
is quite difficult to maintain the mixing scripts for both the
native and cross build. Therefore, I will narrow the focus of this
project to the native builds.

This is the first commit for this drastic change in direction.