Added support of i686-pc-cygwin.
To maintain the consistency between GCC versions and `hack-gcc-specs`
scripts, the build script uses the checksum of `gcc -dumpspecs` to identify
the corresponding `hack-gcc-specs`.

Explicitly prohibits `-fmudflap` and `-fmudflapth` options for GCC on
i686-pc-cygwin because they are not available on i686-pc-cygwin.

Added support of MPICH2 on i686-pc-cygwin.

Refactored `boost/toolset.sh`.

Added and modified `user-config.jam` and `site-config.jam` for
i686-pc-cygwin.

Fixed build script for Boost on i686-pc-cygwin.

Explicitly prohibits the build of Valgrind on i686-pc-cygwin.